more references moved from user_id to identity_id. events left aside
