+++ backend/app/core/webhooks/events.py
@@ -63,7 +63,7 @@ class DataProductRoleAssignmentEvent(V2Event):
 
     id: UUID
     data_product_id: UUID
-    user_id: UUID
+    user_id: UUID # TODO change to indetity_id
 
     @classmethod
     def event_type(cls) -> str:
